requirements.txt
python-telegram-bot[job-queue]==22.5
portalsmp
pyrogram>=2.0.0
python-dotenv==1.0.0
TgCrypto
